pyproject.toml
[project]
name = "ai-daily-journal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "google-cloud-secret-manager>=2.27.0",
    "google-genai>=1.70.0",
    "python-dotenv>=1.2.2",
]
